+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Fixed
+- Close toast notification after refresh ([#136](https://github.com/scm-manager/scm-review-plugin/pull/136))
+- Update comment content on refresh ([#136](https://github.com/scm-manager/scm-review-plugin/pull/136))
+
 ## 2.9.1 - 2021-05-05
 ### Changed
 - Set pull request author as commit author for squash commits via SCMM ([#134](https://github.com/scm-manager/scm-review-plugin/pull/134))